Adding the ability to remember the users last used settings

+++ BitcoinPriceMonitor/PriceMonitor/ITradePriceMonitor.cs
@@ -8,7 +8,7 @@
         double CurrentPrice { get; }
         TradePriceType PriceType { get; set; }
         Currency ConvertToCurrency { get; set; }
-        long Frequency { get; set; }
+        int Frequency { get; set; }
 
         void StartMonitoring();
         void StopMonitoring();
