+++ BitcoinPriceMonitor/Profile/IProfile.cs
@@ -6,7 +6,7 @@
     {
         TradePriceType PriceType { get; set; }
         Currency ConvertToCurrency { get; set; }
-        long Frequency { get; set; }
+        int Frequency { get; set; }
         string MonitorType { get; set; }
     }
 }
